code coverage

+++ .travis-coverage.sh
@@ -4,9 +4,13 @@ echo "" > coverage.txt
 
 for d in $(go list ./... | grep -v vendor); do
     go test -coverprofile=profile.out $d
+    exit_code=$?
+    if [ 0 -ne $exit_code ]; then
+        exit $exit_code
+    fi
     if [ -f profile.out ]; then
         cat profile.out >> coverage.txt
         rm profile.out
     fi
 done
-
+exit 0
